fix: disable the count of active users
The Slack API returns a "limit_required" error when fetching the
presence of 27000+ members.

The server will now use the "conversations.info" endpoint in order to
fetch the total number of users (in the main channel #general). The
count of active users is temporarily disabled, as I don't see any
obvious way to fetch this data. Potential solutions:

- one call per user on the "users.getPresence" endpoint
- use a "presence_query" over the Real Time Messaging API

References:
- https://api.slack.com/methods/users.getPresence
- https://api.slack.com/events/presence_query